Make pre-commit hook more portable between shells
The previous way of doing colours in the shell wasn't portable:

$ cat test.sh
RED='\033[0;31m'
NC='\033[0m'
echo "${RED}hi${NC}"
$ zsh test.sh
hi
$ dash test.sh
hi
$ bash test.sh
\033[0;31mhi\033[0m

The way to portably print anything even slightly complex in shell
scripts is to use printf, as this change does.  We also add a newline
before the summary output and make it bold as well to make it look a
bit more distinct.